+++ packages/batch/src/interfaces/index.ts
@@ -1,3 +1,1 @@
-export type { BatchHooksInterface } from './BatchHooksInterface.js';
-export type { BatchOptionsInterface } from './BatchOptionsInterface.js';
 export type { BatchStatsType } from './BatchStatsType.js';